refactor(rgis-ui): move logic to helper modules
Some UI widgets in the `rgis-ui` crate contained logic that accessed the `GeodesyContext` resource directly. This coupled the presentation logic of the widgets with the application's business logic.

This change refactors the code by moving this logic into dedicated helper functions in a new `ui_helpers` module. Specifically:
- `parse_epsg_input_value` is moved to `ui_helpers::crs_validator`.
- The logic for adding a layer from the library is moved to `ui_helpers::library_helpers`.

This is an incremental step towards a cleaner architecture with better separation of concerns.
